Feat: mypage 포인트 관리기능

+++ src/main/java/org/solo/mypage/service/MypageService.java
@@ -9,11 +9,10 @@ public interface MypageService {
     void insertAssetData(AssetVO assetVO);
     boolean findAssetData(String userID);
     void updateAssetData(AssetVO assetData);
-
     void updateMember(MemberVO memberVO);
 
     int getPoint(String kakaoId);
     boolean withdrawPoints(String kakaoId, int point);
-    boolean updateCash(String userId, double cashAmount);
+    boolean updateCash(String userId, int cashAmount);
 
 }
